Update shipping label in order summary
Changed the shipping label from "Shipping" to "Standard Shipping" in the order summary. The shipping cost is still not displayed.
[skip gpt_engineer]
