+++ src/components/checkout/PaymentForm.tsx
@@ -1,14 +1,13 @@
 import { useState, useEffect } from "react";
-import { useNavigate } from "react-router-dom";
 import { PaymentElement, useStripe, useElements } from "@stripe/react-stripe-js";
-import { Button } from "@/components/ui/button";
+import { useNavigate } from "react-router-dom";
 import { useToast } from "@/components/ui/use-toast";
 import { useCart } from "@/contexts/CartContext";
 import { useAuth } from "@/contexts/AuthContext";
-import { formatCurrency } from "@/lib/utils";
-import { supabase } from "@/integrations/supabase/client";
 import { trackCheckoutStep } from "@/lib/analytics/ecommerce";
 import { createOrder } from "@/lib/utils/paymentUtils";
+import { PaymentFormButton } from "./payment/PaymentFormButton";
+import { useCreateSampleRequest } from "./payment/useCreateSampleRequest";
 
 interface PaymentFormProps {
   clientSecret: string;
@@ -24,96 +23,11 @@ const PaymentForm = ({ clientSecret, total, shippingCost }: PaymentFormProps) =>
   const { items, clearCart } = useCart();
   const { user } = useAuth();
   const [isProcessing, setIsProcessing] = useState(false);
+  const { createRequest } = useCreateSampleRequest();
 
   useEffect(() => {
     trackCheckoutStep(3, items, { total, shipping_cost: shippingCost });
   }, [items, total, shippingCost]);
-
-  const validateZipCode = (zip: string | null) => {
-    if (!zip) return false;
-    // Basic ZIP code validation for Brazil (CEP) and US formats
-    const brZipRegex = /^\d{5}-?\d{3}$/;
-    const usZipRegex = /^\d{5}(-\d{4})?$/;
-    return brZipRegex.test(zip) || usZipRegex.test(zip);
-  };
-
-  const createSampleRequest = async () => {
-    if (!user?.id || !user?.email) throw new Error("User not authenticated");
-
-    const shippingAddress = localStorage.getItem('shipping_address') || '';
-    const shippingCity = localStorage.getItem('shipping_city') || '';
-    const shippingState = localStorage.getItem('shipping_state') || '';
-    const shippingZip = localStorage.getItem('shipping_zip') || '';
-    const firstName = localStorage.getItem('firstName') || '';
-    const lastName = localStorage.getItem('lastName') || '';
-    const phone = localStorage.getItem('phone') || '';
-    const useSameForBilling = localStorage.getItem('useSameForBilling') === 'true';
-
-    // Validate ZIP code
-    if (!validateZipCode(shippingZip)) {
-      throw new Error("Invalid shipping ZIP code format");
-    }
-
-    const billingAddress = useSameForBilling 
-      ? shippingAddress 
-      : localStorage.getItem('billing_address') || '';
-    const billingCity = useSameForBilling 
-      ? shippingCity 
-      : localStorage.getItem('billing_city') || '';
-    const billingState = useSameForBilling 
-      ? shippingState 
-      : localStorage.getItem('billing_state') || '';
-    const billingZip = useSameForBilling 
-      ? shippingZip 
-      : localStorage.getItem('billing_zip') || '';
-
-    // Validate billing ZIP if different
-    if (!useSameForBilling && !validateZipCode(billingZip)) {
-      throw new Error("Invalid billing ZIP code format");
-    }
-
-    const subtotal = items.reduce((sum, item) => sum + (item.from_price * (item.quantity || 1)), 0);
-
-    const { data: sampleRequest, error: sampleRequestError } = await supabase
-      .from('sample_requests')
-      .insert({
-        user_id: user.id,
-        status: 'pending',
-        shipping_address: shippingAddress,
-        shipping_city: shippingCity,
-        shipping_state: shippingState,
-        shipping_zip: shippingZip,
-        billing_address: billingAddress,
-        billing_city: billingCity,
-        billing_state: billingState,
-        billing_zip: billingZip,
-        first_name: firstName,
-        last_name: lastName,
-        payment_method: 'credit_card',
-        shipping_cost: shippingCost,
-        subtotal: subtotal,
-        total: total
-      })
-      .select()
-      .single();
-
-    if (sampleRequestError) throw sampleRequestError;
-
-    const sampleRequestProducts = items.map(item => ({
-      sample_request_id: sampleRequest.id,
-      product_id: item.id,
-      quantity: item.quantity || 1,
-      unit_price: item.from_price
-    }));
-
-    const { error: productsError } = await supabase
-      .from('sample_request_products')
-      .insert(sampleRequestProducts);
-
-    if (productsError) throw productsError;
-
-    return sampleRequest.id;
-  };
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
@@ -125,7 +39,14 @@ const PaymentForm = ({ clientSecret, total, shippingCost }: PaymentFormProps) =>
     setIsProcessing(true);
 
     try {
-      const orderId = await createSampleRequest();
+      const subtotal = items.reduce((sum, item) => sum + (item.from_price * (item.quantity || 1)), 0);
+      const orderId = await createRequest({
+        userId: user.id,
+        items,
+        shippingCost,
+        subtotal,
+        total
+      });
 
       const { error: paymentError, paymentIntent } = await stripe.confirmPayment({
         elements,
@@ -188,13 +109,11 @@ const PaymentForm = ({ clientSecret, total, shippingCost }: PaymentFormProps) =>
       <div className="space-y-4">
         <PaymentElement />
       </div>
-      <Button 
-        type="submit" 
-        disabled={!stripe || isProcessing}
-        className="w-full bg-primary hover:bg-primary-dark"
-      >
-        {isProcessing ? "Processando..." : `Pagar ${formatCurrency(total)}`}
-      </Button>
+      <PaymentFormButton
+        isProcessing={isProcessing}
+        isDisabled={!stripe || isProcessing}
+        total={total}
+      />
     </form>
   );
 };
